Main UI completed

+++ app/src/main/java/dz/akram/bensalem/edvorarideapp/data/Ride.kt
@@ -5,7 +5,7 @@ data class Ride(
     val origin_station_code: Int,
     val station_path: List<Int>,
     val destination_station_code: Int,
-    val date: Int,
+    val date: Long,
     val map_url: String,
     val state: String,
     val city: String
